+++ packages/go_router/lib/src/route_data.dart
@@ -645,3 +645,32 @@ class NoOpPage extends Page<void> {
   Route<void> createRoute(BuildContext context) =>
       throw UnsupportedError('Should never be called');
 }
+
+/// Annotation to override the URI name for a route parameter.
+@Target({TargetKind.parameter})
+class TypedQueryParameter {
+  /// Annotation to override the URI name for a route parameter.
+  const TypedQueryParameter({this.name});
+
+  /// The name of the parameter in the URI.
+  ///
+  /// If `null`, the kebab-case version of the parameter name will be used.
+  ///
+  /// For example:
+  /// ```dart
+  /// class MyRoute extends GoRouteData with $MyRoute {
+  ///   const MyRoute({
+  ///     @TypedQueryParameter(name: 'custom_name') this.myParameter,
+  ///   });
+  ///  final String myParameter;
+  /// }
+  /// ```
+  ///
+  /// This will result in a route that matches
+  /// `/my-route?custom_name=some_value` instead of the default
+  /// `/my-route?my-parameter=some_value`.
+  ///
+  /// It is escaped to be URL-safe. For example `'field with space'` will
+  /// generate a query parameter named `'field+with+space'`.
+  final String? name;
+}