+++ packages/url_launcher/url_launcher_ios/example/lib/main.dart
@@ -35,100 +35,58 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
+  bool _hasCallSupport = false;
   Future<void>? _launched;
   String _phone = '';
 
+  @override
+  void initState() {
+    super.initState();
+    // Check for phone call support.
+    final UrlLauncherPlatform launcher = UrlLauncherPlatform.instance;
+    launcher.canLaunch('tel://123').then((bool result) {
+      if (!mounted) {
+        return;
+      }
+      setState(() {
+        _hasCallSupport = result;
+      });
+    });
+  }
+
   Future<void> _launchInBrowser(String url) async {
     final UrlLauncherPlatform launcher = UrlLauncherPlatform.instance;
-    if (await launcher.canLaunch(url)) {
-      await launcher.launch(
-        url,
-        useSafariVC: false,
-        useWebView: false,
-        enableJavaScript: false,
-        enableDomStorage: false,
-        universalLinksOnly: false,
-        headers: <String, String>{'my_header_key': 'my_header_value'},
-      );
-    } else {
+    if (!await launcher.launchUrl(
+      url,
+      const LaunchOptions(mode: PreferredLaunchMode.externalApplication),
+    )) {
       throw Exception('Could not launch $url');
     }
   }
 
   Future<void> _launchInWebViewOrVC(String url) async {
     final UrlLauncherPlatform launcher = UrlLauncherPlatform.instance;
-    if (await launcher.canLaunch(url)) {
-      await launcher.launch(
-        url,
-        useSafariVC: true,
-        useWebView: true,
-        enableJavaScript: false,
-        enableDomStorage: false,
-        universalLinksOnly: false,
-        headers: <String, String>{'my_header_key': 'my_header_value'},
-      );
-    } else {
-      throw Exception('Could not launch $url');
-    }
-  }
-
-  Future<void> _launchInWebViewWithJavaScript(String url) async {
-    final UrlLauncherPlatform launcher = UrlLauncherPlatform.instance;
-    if (await launcher.canLaunch(url)) {
-      await launcher.launch(
-        url,
-        useSafariVC: true,
-        useWebView: true,
-        enableJavaScript: true,
-        enableDomStorage: false,
-        universalLinksOnly: false,
-        headers: <String, String>{},
-      );
-    } else {
-      throw Exception('Could not launch $url');
-    }
-  }
-
-  Future<void> _launchInWebViewWithDomStorage(String url) async {
-    final UrlLauncherPlatform launcher = UrlLauncherPlatform.instance;
-    if (await launcher.canLaunch(url)) {
-      await launcher.launch(
-        url,
-        useSafariVC: true,
-        useWebView: true,
-        enableJavaScript: false,
-        enableDomStorage: true,
-        universalLinksOnly: false,
-        headers: <String, String>{},
-      );
-    } else {
+    if (!await launcher.launchUrl(
+      url,
+      const LaunchOptions(mode: PreferredLaunchMode.inAppBrowserView),
+    )) {
       throw Exception('Could not launch $url');
     }
   }
 
   Future<void> _launchUniversalLinkIos(String url) async {
     final UrlLauncherPlatform launcher = UrlLauncherPlatform.instance;
-    if (await launcher.canLaunch(url)) {
-      final bool nativeAppLaunchSucceeded = await launcher.launch(
+    final bool nativeAppLaunchSucceeded = await launcher.launchUrl(
+      url,
+      const LaunchOptions(
+        mode: PreferredLaunchMode.externalNonBrowserApplication,
+      ),
+    );
+    if (!nativeAppLaunchSucceeded) {
+      await launcher.launchUrl(
         url,
-        useSafariVC: false,
-        useWebView: false,
-        enableJavaScript: false,
-        enableDomStorage: false,
-        universalLinksOnly: true,
-        headers: <String, String>{},
+        const LaunchOptions(mode: PreferredLaunchMode.inAppBrowserView),
       );
-      if (!nativeAppLaunchSucceeded) {
-        await launcher.launch(
-          url,
-          useSafariVC: true,
-          useWebView: true,
-          enableJavaScript: false,
-          enableDomStorage: false,
-          universalLinksOnly: true,
-          headers: <String, String>{},
-        );
-      }
     }
   }
 
@@ -142,17 +100,7 @@ class _MyHomePageState extends State<MyHomePage> {
 
   Future<void> _makePhoneCall(String url) async {
     final UrlLauncherPlatform launcher = UrlLauncherPlatform.instance;
-    if (await launcher.canLaunch(url)) {
-      await launcher.launch(
-        url,
-        useSafariVC: false,
-        useWebView: false,
-        enableJavaScript: false,
-        enableDomStorage: false,
-        universalLinksOnly: true,
-        headers: <String, String>{},
-      );
-    } else {
+    if (!await launcher.launchUrl(url, const LaunchOptions())) {
       throw Exception('Could not launch $url');
     }
   }
@@ -177,10 +125,14 @@ class _MyHomePageState extends State<MyHomePage> {
                 ),
               ),
               ElevatedButton(
-                onPressed: () => setState(() {
-                  _launched = _makePhoneCall('tel:$_phone');
-                }),
-                child: const Text('Make phone call'),
+                onPressed: _hasCallSupport
+                    ? () => setState(() {
+                        _launched = _makePhoneCall('tel:$_phone');
+                      })
+                    : null,
+                child: _hasCallSupport
+                    ? const Text('Make phone call')
+                    : const Text('Calling not supported'),
               ),
               const Padding(
                 padding: EdgeInsets.all(16.0),
@@ -198,18 +150,6 @@ class _MyHomePageState extends State<MyHomePage> {
                   _launched = _launchInWebViewOrVC(toLaunch);
                 }),
                 child: const Text('Launch in app'),
-              ),
-              ElevatedButton(
-                onPressed: () => setState(() {
-                  _launched = _launchInWebViewWithJavaScript(toLaunch);
-                }),
-                child: const Text('Launch in app(JavaScript ON)'),
-              ),
-              ElevatedButton(
-                onPressed: () => setState(() {
-                  _launched = _launchInWebViewWithDomStorage(toLaunch);
-                }),
-                child: const Text('Launch in app(DOM storage ON)'),
               ),
               const Padding(padding: EdgeInsets.all(16.0)),
               ElevatedButton(