+++ packages/video_player/video_player_avfoundation/darwin/video_player_avfoundation/Sources/video_player_avfoundation/AVAssetTrackUtils.m
@@ -2,11 +2,9 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#import <AVFoundation/AVFoundation.h>
+@import AVFoundation;
 
-CGAffineTransform FVPGetStandardizedTransformForTrack(AVAssetTrack *track) {
-  CGAffineTransform t = track.preferredTransform;
-  CGSize size = track.naturalSize;
+CGAffineTransform FVPGetStandardizedTrackTransform(CGAffineTransform t, CGSize size) {
   // Each case of control flows corresponds to a specific
   // `UIImageOrientation`, with 8 cases in total.
   if (t.a == 1 && t.b == 0 && t.c == 0 && t.d == 1) {