+++ packages/video_player/video_player_avfoundation/darwin/video_player_avfoundation/Sources/video_player_avfoundation/include/video_player_avfoundation/FVPVideoPlayer_Internal.h
@@ -2,7 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#import <AVFoundation/AVFoundation.h>
+@import AVFoundation;
+
 #import "FVPAVFactory.h"
 #import "FVPVideoEventListener.h"
 #import "FVPVideoPlayer.h"
@@ -12,8 +13,8 @@ NS_ASSUME_NONNULL_BEGIN
 
 /// Interface intended for use by subclasses, but not other callers.
 @interface FVPVideoPlayer ()
-/// The AVPlayerItemVideoOutput associated with this video player.
-@property(nonatomic, readonly) AVPlayerItemVideoOutput *videoOutput;
+/// The pixel buffer source associated with this video player.
+@property(nonatomic, readonly) NSObject<FVPPixelBufferSource> *pixelBufferSource;
 /// The view provider, to obtain view information from.
 @property(nonatomic, readonly, nullable) NSObject<FVPViewProvider> *viewProvider;
 /// The preferred transform for the video. It can be used to handle the rotation of the video.