+++ packages/go_router_builder/CHANGELOG.md
@@ -1,6 +1,7 @@
-## NEXT
+## 4.2.0
 
-* Adds support for `onEnter` method in `GoRouteData` classes to implement navigation guards.
+- Adds supports for `TypedQueryParameter` annotation.
+  Its `name` parameter allows specifying a different name for the query parameter than the field name. The name is escaped to be URL-safe. For example `@TypedQueryParameter(name: 'field with space')` will generate a query parameter named `field+with+space`.
 
 ## 4.1.3